Refinement for π-Cation interactions fixed and improved
For protein-protein interactions, HIS residues, which provide charge and a ring at the same time, can also be present on the ligand side, making the refinement step needed there as well. Also, refinement was improved to make sure it is indeed the same HIS residue that is already also forming a π-stacking interaction.